testing: create base test class and refactor test_copy test suite.
